+++ oalg-base/src/OAlg/Data/Boolean/Definition.hs
@@ -14,11 +14,8 @@
 -- 
 -- boolean structure for multivalent logic. 
 module OAlg.Data.Boolean.Definition
-  ( -- * Logical Operators
-    Logical(..), Erasable(..)
-    
-    -- * Boolean
-  , Boolean(..)
+  ( -- * Boolean
+    Boolean(..)
 
     -- * Bool
   , B.Bool(..), B.otherwise
@@ -31,48 +28,8 @@ module OAlg.Data.Boolean.Definition
 import Prelude as P hiding (not,(||),(&&),or,and)
 import qualified Data.Bool as B
 
+import OAlg.Data.Logical
 import OAlg.Data.Opposite
-
---------------------------------------------------------------------------------
--- Lagical -
-
--- | logical structures admitting a general definition for /disjunctions/ and /conjunctions/.
-class Logical a where
-  infixr 2 ||
-  -- | disjunction
-  (||) :: a -> a -> a
-  
-  infixr 3 &&  
-  -- | conjunction
-  (&&) :: a -> a -> a
-
-instance Logical Bool where
-  (||)  = (B.||)
-  (&&)  = (B.&&)
-
-instance Logical a => Logical (Op a) where
-  Op a || Op b = Op (a && b)
-  Op a && Op b = Op (a || b)
-
---------------------------------------------------------------------------------
--- Erasable -
-
--- | erasor-operator.
-class Erasable a where
-  infixl 4 //
-  -- | difference
-  (//) :: a -> a -> a
-  default (//) :: Boolean a => a -> a -> a
-  a // b = a && not b
-
-instance Erasable Bool
-
-instance Eq x => Erasable [x] where
-  xs // [] = xs
-  [] // _  = []
-  (x:xs) // (y:ys) = case x == y of
-    True  -> xs // ys
-    False -> x : (xs // (y:ys))
 
 --------------------------------------------------------------------------------
 -- Boolean -